[FIX] base: choose latest translation first
When synchronizing translations, the algorithm search for the translated
terms without specifying an order.
As a result, the translation terms are in a non deterministic order and
the first one is chosen which might not be the latest.

During the tests, this bug appeared from times to times during the
rte_translation tour. One way to force it, is to order the translations
by value descending.

With this commit, the translations are ordered by id descending, as a
consequence, the chosen transaltion will always be the latest.

+++ odoo/addons/base/models/ir_translation.py
@@ -469,7 +469,7 @@ class IrTranslation(models.Model):
                 ('type', '=', 'model_terms'),
                 ('name', '=', "%s,%s" % (field.model_name, field.name)),
                 ('res_id', '=', record.id),
-            ])
+            ], order='id DESC')
 
             if not terms:
                 # discard all translations for that field
@@ -494,8 +494,10 @@ class IrTranslation(models.Model):
                         outdated += trans
 
         # process outdated and discarded translations
-        outdated.write({'state': 'to_translate'})
-        discarded.unlink()
+        if outdated:
+            outdated.write({'state': 'to_translate'})
+        if discarded:
+            discarded.unlink()
 
     @api.model
     @tools.ormcache_context('model_name', keys=('lang',))
